Add clarity to the code and project files name
