+++ OC_FrenchGameFactory/OC_FrenchGameFactory/View/HomeScreen/CreateButton.swift
@@ -8,7 +8,9 @@
 import Foundation
 import UIKit
 
-//Create a random color when called on a UIColor object
+
+
+///Create a random color when called on a UIColor object
 extension UIColor {
     static var random: UIColor {
         return UIColor(
@@ -20,16 +22,46 @@ extension UIColor {
     }
 }
 
+
+
+
+
 class CreateButton: UIButton{
+    
+    
+    // Character corresponding to this button instance
     var correspondingCharacter: Character?
-    var correspondingCustomName: UITextView? // Used to display or hide the textView if button is pressed
+    
+    // Used to display or hide the textView corresponding to the character custom name of this button instance when it is pressed
+    var correspondingCustomName: UITextView?
+    
+    // Used for button actions when pressed
     var pressed : Bool?
     
+    
+    
     override init(frame: CGRect) {
+        
+        
+        
         super.init(frame: frame)
+    
+    
+    
     }
 
+    
+    
     required init?(coder aDecoder: NSCoder) {
+    
+        
+        
         super.init(coder: aDecoder)
+    
+    
+    
     }
+
+
+
 }
